Merge Details:

Jonathan Squirawski (3):
      6949b1085d build: add ASDF requirements to build the lib-common
      b28bbcfec2 build: fix ctags definitions for Universal Ctags 5.9
      dec9ec2cb2 iopc: fix compilation with GCC 12.2.1

Conflicts: Build/bulk.mk
 iopc/iopc-parser.c

+++ iopc/iopc-parser.c
@@ -2288,8 +2288,8 @@ parse_function_stmt(iopc_parser_t *pp, qv_t(iopc_attr) *attrs,
     parse_function_desc(pp, IOP_F_ARGS, fun, &arg_chunks, type);
 
     /* XXX we use & to execute both function calls */
-    if ((!parse_function_desc(pp, IOP_F_RES,  fun, &arg_chunks, type))
-     &  (!parse_function_desc(pp, IOP_F_EXN,  fun, &arg_chunks, type)))
+    if ((int)(!parse_function_desc(pp, IOP_F_RES,  fun, &arg_chunks, type))
+     &  (int)(!parse_function_desc(pp, IOP_F_EXN,  fun, &arg_chunks, type)))
     {
         if (!iopc_is_snmp_iface(type)) {
             info_loc("function %s may be a candidate for async-ness",
